+++ VelogOnMobile/Feature/ScrapFolderBottomSheet/View/ScrapFolderBottomSheetView.swift
@@ -23,7 +23,7 @@ final class ScrapFolderBottomSheetView: BaseUIView {
     
     private let titleLabel: UILabel = {
         let label = UILabel()
-        label.text = "폴더 선택"
+        label.text = TextLiterals.bottomSheetText
         label.textAlignment = .center
         label.textColor = .black
         label.font = UIFont(name: "Avenir-Black", size: 18)
@@ -33,7 +33,26 @@ final class ScrapFolderBottomSheetView: BaseUIView {
     private let cancelButton: UIButton = {
         let button = UIButton()
         button.setImage(UIImage(systemName: "xmark"), for: .normal)
+        button.tintColor = .darkGrayColor
         return button
+    }()
+    
+    private lazy var newFolderButton: UIButton = {
+        let button = UIButton()
+        button.setImage(ImageLiterals.plusFolder, for: .normal)
+        button.backgroundColor = .white
+        button.layer.cornerRadius = 4
+        button.layer.masksToBounds = true
+        return button
+    }()
+    
+    private let folderTitleLabel: UILabel = {
+        let label = UILabel()
+        label.text = TextLiterals
+            .makeNewFolderButtonText
+        label.textColor = .black
+        label.font = UIFont(name: "Avenir-Black", size: 15)
+        return label
     }()
     
     let folderTableView: UITableView = {
@@ -42,10 +61,9 @@ final class ScrapFolderBottomSheetView: BaseUIView {
         tableView.backgroundColor = .white
         tableView.showsVerticalScrollIndicator = false
         tableView.rowHeight = 46
-        tableView.sectionHeaderHeight = 49
+        tableView.sectionHeaderHeight = 0
         return tableView
     }()
-    
     
     // MARK: - Functions
     
@@ -58,6 +76,8 @@ final class ScrapFolderBottomSheetView: BaseUIView {
         bottomSheetView.addSubviews(
             titleLabel,
             cancelButton,
+            newFolderButton,
+            folderTitleLabel,
             folderTableView
         )
         
@@ -67,7 +87,7 @@ final class ScrapFolderBottomSheetView: BaseUIView {
         }
         
         titleLabel.snp.makeConstraints {
-            $0.top.equalToSuperview().inset(14)
+            $0.top.equalToSuperview().inset(12)
             $0.centerX.equalToSuperview()
         }
         
@@ -76,8 +96,19 @@ final class ScrapFolderBottomSheetView: BaseUIView {
             $0.top.trailing.equalToSuperview().inset(18)
         }
         
+        newFolderButton.snp.makeConstraints {
+            $0.top.equalToSuperview().offset(60)
+            $0.leading.equalToSuperview().inset(20)
+            $0.height.width.equalTo(26)
+        }
+        
+        folderTitleLabel.snp.makeConstraints {
+            $0.centerY.equalTo(newFolderButton)
+            $0.leading.equalTo(newFolderButton.snp.trailing).offset(10)
+        }
+    
         folderTableView.snp.makeConstraints {
-            $0.top.equalTo(titleLabel.snp.bottom)
+            $0.top.equalTo(newFolderButton.snp.bottom)
             $0.leading.trailing.bottom.equalToSuperview()
         }
     }
